mypy - black - make sure tests mock properly
