+++ tests/test_client.py
@@ -1,9 +1,39 @@
+from __future__ import annotations
+
+
 from auth0.v3.exceptions import Auth0Error
 import pytest
+from typing import Any
 
 from unittest.mock import patch
 from sdk import SDK, Endpoints, Auth0Conf
-from sdk.authentication import FakeAuth0BadAuthentication, FakeAuth0GoodAuthentication
+from sdk.authentication import TokenProvider
+
+
+class FakeAuth0GoodAuthentication(TokenProvider):
+    def _query_token(self) -> dict[str, Any]:
+        return {
+            "access_token": "some_token",
+            "id_token": "id_token",
+            "scope": "openid profile email",
+            "expires_in": 86400,
+            "token_type": "Bearer",
+        }
+
+
+class FakeAuth0BadAuthentication(TokenProvider):
+    def __init__(self, *args: tuple, **kwags: dict):
+        """The arguments are not important.
+        What's important is that the init raise the error below.
+        """
+        self._query_token()
+
+    def _query_token(self) -> dict[str, Any]:
+        """The lib raises a Auth0Error, but I raise a ValueError in order
+        not to confused the tests. If a Auth0Error is raised, it's
+        because we didn't mock properly.
+        """
+        raise ValueError()
 
 
 @patch("sdk.client.Auth0TokenProvider", FakeAuth0GoodAuthentication)
@@ -22,10 +52,6 @@ class TestAuthSuccess:
     def test_authentication_success(self):
         SDK(group_id=self.group_id, username=self.username, password=self.password)
 
-    def test_authentication_no_credentials_provided(self):
-        with pytest.raises(ValueError):
-            SDK(group_id=self.group_id)
-
     def test_good_token_provider(self):
         SDK(group_id=self.group_id, token_provider=FakeAuth0GoodAuthentication)
 
@@ -38,18 +64,6 @@ class TestAuthSuccess:
         )
         assert sdk._client.endpoints.core == self.new_core_endpoint
 
-    def test_bad_endpoints(self):
-        with pytest.raises(TypeError):
-            SDK(
-                group_id=self.group_id,
-                username=self.username,
-                password=self.password,
-                endpoints={
-                    "core": "",
-                    "account": "",
-                },
-            )
-
     def test_correct_new_auth0(self):
         new_auth0 = Auth0Conf(domain="new_domain")
         SDK(
@@ -58,15 +72,6 @@ class TestAuthSuccess:
             password=self.password,
             auth0=new_auth0,
         )
-
-    def test_bad_auth0(self):
-        with pytest.raises(TypeError):
-            SDK(
-                group_id=self.group_id,
-                username=self.username,
-                password=self.password,
-                auth0="",
-            )
 
 
 @patch("sdk.client.Auth0TokenProvider", FakeAuth0BadAuthentication)
@@ -78,22 +83,17 @@ class TestAuthFailure:
     no_password = ""
 
     @patch("sdk.client.getpass")
-    def test_module_getpass_no_password(self, getpass):
-        getpass.return_value = self.no_password
+    def test_module_getpass_bad_password(self, getpass):
+        getpass.return_value = self.password
 
         with pytest.raises(ValueError):
             SDK(group_id=self.group_id, username=self.username)
 
         getpass.assert_called_once()
 
-    @patch("sdk.client.getpass")
-    def test_module_getpass_bad_password(self, getpass):
-        getpass.return_value = self.password
-
-        with pytest.raises(Auth0Error):
-            SDK(group_id=self.group_id, username=self.username)
-
-        getpass.assert_called_once()
+    def test_module_bad_password(self):
+        with pytest.raises(ValueError):
+            SDK(group_id=self.group_id, username=self.username, password=self.password)
 
 
 class TestAuthInvalidClient:
@@ -121,10 +121,40 @@ class TestAuthInvalidClient:
                 password=self.no_password,
             )
 
-    def test_module_bad_password(self):
-        with pytest.raises(Auth0Error):
-            SDK(group_id=self.group_id, username=self.username, password=self.password)
+    @patch("sdk.client.getpass")
+    def test_module_getpass_no_password(self, getpass):
+        getpass.return_value = self.no_password
+
+        with pytest.raises(ValueError):
+            SDK(group_id=self.group_id, username=self.username)
+
+        getpass.assert_called_once()
 
     def test_bad_token_provider(self):
         with pytest.raises(TypeError):
             SDK(group_id=self.group_id, token_provider="token")
+
+    def test_bad_auth0(self):
+        with pytest.raises(TypeError):
+            SDK(
+                group_id=self.group_id,
+                username=self.username,
+                password=self.password,
+                auth0="",
+            )
+
+    def test_authentication_no_credentials_provided(self):
+        with pytest.raises(ValueError):
+            SDK(group_id=self.group_id)
+
+    def test_bad_endpoints(self):
+        with pytest.raises(TypeError):
+            SDK(
+                group_id=self.group_id,
+                username=self.username,
+                password=self.password,
+                endpoints={
+                    "core": "",
+                    "account": "",
+                },
+            )
